Big update with new functions
clickSenator function that show more details for each senator when a row is clicked.

Sort of the data before the senators are added to the table. It groups all senators by party by default

Specific message displayed if a party has no leaders.

Code that integrates all the corresponding options for each filter. (e.g: all states for State filter ) in the dropdowns

Other small updates and bug fixes